refactor: Major overhaul to audio clip retention methods, age and usage based policy options
