+++ internal/conf/defaults.go
@@ -40,10 +40,10 @@ func setDefaultConfig() {
 
 	viper.SetDefault("realtime.audio.export.retention.enabled", true)
 	viper.SetDefault("realtime.audio.export.retention.debug", false)
-	viper.SetDefault("realtime.audio.export.retention.mode", "priority")
-	viper.SetDefault("realtime.audio.export.retention.diskusagelimit", "80%")
-	viper.SetDefault("realtime.audio.export.retention.minClipsPerSpecies", 10)
-	viper.SetDefault("realtime.audio.export.retention.minEvictionHours", 0)
+	viper.SetDefault("realtime.audio.export.retention.policy", "use")
+	viper.SetDefault("realtime.audio.export.retention.maxusage", "80%")
+	viper.SetDefault("realtime.audio.export.retention.maxage", "30d")
+	viper.SetDefault("realtime.audio.export.retention.minclips", 10)
 
 	viper.SetDefault("realtime.log.enabled", false)
 	viper.SetDefault("realtime.log.path", "birdnet.txt")
